implement validation

+++ app/Services/ProblemService/ProblemTableActionService.php
@@ -2,20 +2,29 @@
 
 namespace App\Services\ProblemService;
 
+use App\Enums\ProblemEnum;
+use App\Enums\RoleAndPermissionEnum;
 use App\Interfaces\TableActionInterface;
+use App\Services\RoleAndPermission\PermissionService;
 use App\Services\UserService\UserService;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\Rule;
 use Mockery\Exception;
 
 class ProblemTableActionService implements TableActionInterface
 {
 
     private ProblemService $problemService;
+    private PermissionService $permissionService;
+    private $validationMessage;
 
     public function __construct()
     {
         $this->problemService = new ProblemService();
+        $this->permissionService = new PermissionService();
+        $this->validationMessage = __('app.validation');
     }
 
 
@@ -28,34 +37,61 @@ class ProblemTableActionService implements TableActionInterface
     {
         try {
             DB::beginTransaction();
+            $isPermitted = $this->permissionService->isHasPermission(permission: RoleAndPermissionEnum::PERMISSION_MANAGE_PROBLEM, crudKey: RoleAndPermissionEnum::CREATE);
+            if (!$isPermitted) {
+                throw new Exception(__('app.response.error.permission_denied'));
+            }
+            $data = Validator::make($data, [
+                'problem' => 'required|string',
+            ], $this->validationMessage)->validate();
+
             $problem = $this->problemService->store($data);
             DB::commit();
         } catch (Exception $exception) {
             DB::rollBack();
-            throw new Exception( $exception->getMessage() . ' ' . $exception->getLine());
+            throw new Exception($exception->getMessage() . ' ' . $exception->getLine());
         }
-        return ['data' => $problem, 'message' => 'Create success'];
+        return ['data' => $problem, 'message' => __('app.response.success.store')];
     }
 
     public function update(string $id, array $data)
     {
         try {
             DB::beginTransaction();
+            $isPermitted = $this->permissionService->isHasPermission(permission: RoleAndPermissionEnum::PERMISSION_MANAGE_PROBLEM, crudKey: RoleAndPermissionEnum::UPDATE);
+            if (!$isPermitted) {
+                throw new Exception(__('app.response.error.permission_denied'));
+            }
+            $data = Validator::make($data, [
+                'problem' => 'string|required',
+                'note' => [
+                    'string',
+                    Rule::requiredIf(function () use ($data) {
+                        return isset($data['status']) && ($data['status'] != ProblemEnum::SEND || $data['status'] != ProblemEnum::PROCESS);
+                    }),
+                ],
+                'status' => 'string|nullable',
+            ], $this->validationMessage)->validate();
+
             $problem = $this->problemService->update($id, $data);
             DB::commit();
         } catch (Exception $exception) {
             DB::rollBack();
-            throw new Exception( $exception->getMessage() . ' ' . $exception->getLine());
+            throw new Exception($exception->getMessage() . ' ' . $exception->getLine());
         }
-        return ['data' => $problem, 'message' => 'Update success'];
+        return ['data' => $problem, 'message' => __('app.response.success.update')];
     }
 
     public function find($id)
     {
         try {
+            Validator::make(['id' => $id], [
+                'id' => 'required|string',
+            ], $this->validationMessage)->validate();
+
             $problem = $this->problemService->find($id);
         } catch (Exception $exception) {
-            throw new Exception( $exception->getMessage() . ' ' . $exception->getLine());
+            throw new Exception($exception->getMessage() . ' ' . $exception->getLine());
         }
         return $problem;
 
@@ -69,14 +105,21 @@ class ProblemTableActionService implements TableActionInterface
     public function delete($id)
     {
         try {
+            $isPermitted = $this->permissionService->isHasPermission(permission: RoleAndPermissionEnum::PERMISSION_MANAGE_PROBLEM, crudKey: RoleAndPermissionEnum::DELETE);
+            if (!$isPermitted) {
+                throw new Exception(__('app.response.error.permission_denied'));
+            }
+            Validator::make(['id' => $id], [
+                'id' => 'required|string',
+            ], $this->validationMessage)->validate();
             DB::beginTransaction();
             $problem = $this->problemService->find($id);
             $this->problemService->delete($id);
             DB::commit();
         } catch (Exception $exception) {
             DB::rollBack();
-            throw new Exception( $exception->getMessage() . ' ' . $exception->getLine());
+            throw new Exception($exception->getMessage() . ' ' . $exception->getLine());
         }
-        return ['data' => $problem, 'message' => 'Delete success'];
+        return ['data' => $problem, 'message' => __('app.response.success.delete')];
     }
 }
